requirements.txt
transformers==4.37.2
tweepy==4.14.0
python-dotenv==1.0.1
schedule==1.2.1
torch>=2.0.0
numpy==1.24.3
tqdm==4.66.1
pandas==2.1.4
SQLAlchemy==2.0.25